Added Auth And Contract Functions and Date Formatter

+++ database/migrations/2024_04_17_113751_create_contracts_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('contracts', function (Blueprint $table) {
             $table->id();
-            $table->float('price');
+            $table->decimal('price', 12);
             $table->date('start_date');
             $table->date('end_date');
             $table->float('duration');
